requirements.txt
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
spacy>=3.7.0
scikit-learn>=1.3.0
nltk>=3.8.0
python-dateutil>=2.8.0
